+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Fixed
+
+- map `$SpiSlave` into `SpiSlave` struct in `spi!` macro [#635]
+
+[#635]: https://github.com/stm32-rs/stm32f4xx-hal/pull/635
+
 ## [v0.16.0] - 2023-05-07
 
 ### Changed